Alterado tipo dos ids'

+++ backend/DTO/Consulta/ConsultaPostDTO.cs
@@ -5,9 +5,9 @@ namespace backend.DTO
     public class ConsultaPostDTO
     {
         [Required(ErrorMessage = "O campo {0} é obrigatório.")]
-        public int PacienteId { get; set; }
+        public long PacienteId { get; set; }
         [Required(ErrorMessage = "O campo {0} é obrigatório.")]
-        public int DentistaId { get; set; }
+        public long DentistaId { get; set; }
         [Required(ErrorMessage = "O campo {0} é obrigatório.")]
         public DateTime DataHora { get; set; }
         public ICollection<ProocedimentoConsultaPostDTO>? ListaProcedimentos { get; set; }
