+++ backend/Models/Consulta.cs
@@ -6,10 +6,10 @@ namespace SistemaOdontologico.Models;
 public class Consulta
 {
     [Key]
-    public int Id { get; set; }
-    public int PacienteId { get; set; }
+    public long Id { get; set; }
+    public long PacienteId { get; set; }
     public Paciente? Paciente { get; set; }
-    public int DentistaId { get; set; }
+    public long DentistaId { get; set; }
     public Dentista? Dentista { get; set; }
     public DateTime DataHora { get; set; }
     public ICollection<ProcedimentoConsulta>? Procedimentos { get; set; } = [];
